Don't use instaceof to detect dates.
The reason for this change is that instanceof breaks across node sandboxes.

Ultimately it would be better to move to a function that detects if an object
has the methods needed to manage dates (getUTC*/setUTC*) so that users may
write own date classes/subclasses. Will cross that bridge when I find it.